Remove apparently unnecessary CTLL_FIXED_STRING definition
This seems to not be used since a long time and its actual definition
was broken by the changes of c967f68 (support for clang12, 2021-03-25)
as CTLL_CNTTP_COMPILER_CHECK wasn't defined in this header, which was
included before utilities.hpp defining this symbol, so it was actually
always defined as "const auto &".

Incidentally, this fixes (correct) gcc -Wundef warning about checking
CTLL_CNTTP_COMPILER_CHECK before it is defined.