Disable `.pc` generation by default
The problem with generating a pkgconfig file via CMake is that it does
not fit the model of CMake. The `prefix` variable has to be absolute in
the .pc file, however CMake deals with paths relative to the prefix. The
final prefix can change at all times, via `CMAKE_INSTALL_PREFIX`,
`--prefix` parameter during install time, at CPack time and the
resulting deb/rpm files can be extracted anywhere as well. If one needs
a pkgconfig file for packaging purposes, both deb and rpm CPack
generators allow the user to add post-install scripts that can generate
it if necessary.