cmake: claim knowledge of CMake 3.29
This is true because the code considers 3.29 in the next block. Of
particular importance, this sets CMP0115 which scans C++20 sources by
default. This allows tests to use `add_executable(t uses_module.cpp)`
without having to explicitly say "I want to use `import ctre;`".

In the non-module build, explicitly turn off scanning since it is not
used.